pyproject.toml
[project]
name = "exaqube-task"
version = "0.1.0"
description = ""
authors = [
    {name = "venkat"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph (>=0.3.29,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "toolbox-langchain (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

